cambion OrderSummay y borrar puerta de la orden

+++ Model/Order.cs
@@ -11,7 +11,7 @@ namespace Model
         public int Id { get; set; }
         public User User { get; set; }
         public int Quantity { get; set; }
-        public decimal Total { get; set; }
+        public int Total { get; set; }
         public Type Type { get; set; }
         public Status Status { get; set; }
         public DateTime CreationDate { get; set; }
